+++ CHANGELOG.md
@@ -20,10 +20,12 @@ Version 10 of the SDK is compatible with Sentry self-hosted versions 24.4.2 or h
 
 ### Features
 
+- Add Fallback to JavaScript SDK when Native SDK fails to initialize ([#1043](https://github.com/getsentry/sentry-capacitor/pull/1043))
 - Add spotlight integration `spotlightIntegration`. ([#1039](https://github.com/getsentry/sentry-capacitor/pull/1039))
 
 ### Bugfix
 
+- Replays/Logs/Sessions now have the `capacitor` SDK name as the source of the event. ([#1043](https://github.com/getsentry/sentry-capacitor/pull/1043))
 - Sentry Capacitor integrations are now exposed to `@sentry/capacitor` ([#1039](https://github.com/getsentry/sentry-capacitor/pull/1039))
 
 ### Removed APIs
@@ -58,9 +60,9 @@ For more informations, please go to the following link: https://docs.sentry.io/p
 
 ### Dependencies
 
-- Bump JavaScript Sibling SDKs from v9.46.0 to v10.20.0 ([#1013](https://github.com/getsentry/sentry-capacitor/pull/1013))
-  - [changelog](https://github.com/getsentry/sentry-javascript/blob/10.20.0/CHANGELOG.md)
-  - [diff](https://github.com/getsentry/sentry-javascript/compare/9.46.0...10.20.0)
+- Bump JavaScript Sibling SDKs from v9.46.0 to v10.27.0 ([#1013](https://github.com/getsentry/sentry-capacitor/pull/1013), [#1028](https://github.com/getsentry/sentry-capacitor/pull/1028))
+  - [changelog](https://github.com/getsentry/sentry-javascript/blob/10.27.0/CHANGELOG.md)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/9.46.0...10.27.0)
 
 ## 2.4.1